+++ workspaces/ai-integrations/packages/backend/src/index.ts
@@ -82,4 +82,9 @@ backend.add(
     '@red-hat-developer-hub/backstage-plugin-catalog-backend-module-catalog-backend-module-ai-resource-extensions'
   ),
 );
+backend.add(
+  import(
+    '@red-hat-developer-hub/backstage-plugin-kserve-kubeflow-connector-backend'
+  ),
+);
 backend.start();